update AHI HSD to explicitly check if a temporary file has been created prior to attempting deletion

+++ satpy/readers/ahi_hsd.py
@@ -308,7 +308,7 @@ class AHIHSDFileHandler(BaseFileHandler):
         self.calib_mode = calib_mode.upper()
 
     def __del__(self):
-        if (self.is_zipped):
+        if (self.is_zipped and os.path.exists(self.filename)):
             os.remove(self.filename)
 
     @property
